Automatically prepend select type fields with empty values, fix selects usage for relations
